Add new CORS Method 'ACKEE_AUTO_FQDN_ORIGIN'
Previously there were two builtin methods of adding CORS headers to
ACKEE: One could either set a wildcard header (`ACKEE_ALLOW_ORIGIN=*`)
or set a concrete list of domains
(`ACKEE_ALLOW_ORIGIN=domain1.example.com,domain2.example.com`). However
these became very annoying to configure when working with more than a
couple of domains.

This commit proposes adding a new method 'ACKEE_AUTO_FQDN_ORIGIN' to
figure out which domains to add CORS Headers for. Instead of hard-coding
the domains, this mode instead determines them from the list of tracking domains.
Conretely, each domain in the domain list that has a fully qualified
domain name as a title is treated as a valid domain to add a CORS header for.

The implementation is achieved by means of a new 'domainFqNames.js'
utility, which uses the 'is-valid-domain' npm package to filter the list
of domains. Additionally the 'findMatchingOrigin.js' utility is modified
to make use of the above. A second equivalent implementation is added
to 'serverless.js'.

This commit also adds a test for the new functionality, although this
does not test the serverless implementation.